extract the tests of LetInExpander into a designated class
